pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "commit"
version = "0.1.0"
description = "Commit to your learning journey - convert LaTeX notes to intelligent Anki flashcards"
readme = "README.md"
requires-python = ">=3.11"
authors = [
    {name = "Erik Bach Ryhl"}
]
dependencies = [
    "gitpython>=3.1.40",
    "pydantic>=2.5.0",
    "httpx>=0.25.0",
    "pyyaml>=6.0",
    "typer>=0.9.0",
    "rich>=13.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "respx>=0.20.0",
    "genanki>=0.13.0",
]

[project.scripts]
commit = "commit.cli:app"

[tool.setuptools.packages.find]
include = ["commit*"]

